:recycle: move contribution status into its own file
